Seperated functions and tests according to requirements
