+++ javascript/reverseString.js
@@ -1,0 +1,9 @@
+const reverseString = (string) => {
+  try {
+    return string.split('').reverse().join('');
+  } catch (err) {
+    return 'Input valid string';
+  }
+};
+
+export default reverseString;
